requirements.txt
boto3>=1.4.0
PyYAML>=3.11
